# openapi.yaml
openapi: 3.0.3
info:
  title: Treasury Management System API
  version: "1.0.0"
  description: >
    REST API for the TMS Phase 1 MVP - bank accounts, cash position,
    payments, incoming transactions, inter-bank transfers, approvals,
    cash forecasting, reports, administration, and audit logs.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /bank-accounts:
    get:
      tags: [Bank Accounts]
      summary: List bank accounts with computed cash metrics
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/search"
        - in: query
          name: bankId
          schema: { type: string }
        - in: query
          name: currencyCode
          schema: { type: string }
      responses:
        "200": { description: OK }
    post:
      tags: [Bank Accounts]
      summary: Create a bank account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BankAccountInput" }
      responses:
        "201": { description: Created }
  /bank-accounts/{id}:
    get:
      tags: [Bank Accounts]
      summary: Get a bank account by id
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: OK }
    patch:
      tags: [Bank Accounts]
      summary: Update an account, including minimum/target balance configuration
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: OK }
  /bank-accounts/{id}/balance:
    post:
      tags: [Bank Accounts]
      summary: Record a manual balance update for an account
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [balanceDate, closingBalance]
              properties:
                balanceDate: { type: string, format: date }
                closingBalance: { type: number }
                availableBalance: { type: number }
      responses:
        "200": { description: OK }
...
  /payments:
    get:
      tags: [Payments]
      summary: List payments
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/search"
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200": { description: OK }
    post:
      tags: [Payments]
      summary: Create a draft payment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentInput" }
      responses:
        "201": { description: Created }
  /payments/{id}:
    get:
      tags: [Payments]
      summary: Get a payment
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: OK }
  /payments/{id}/submit:
    post:
      tags: [Payments]
      summary: Submit a draft payment for approval
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: OK }
  /payments/{id}/cancel:
    post:
      tags: [Payments]
      summary: Cancel a draft or pending payment
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: OK }
...
  /forecasts/projection:
    get:
      tags: [Forecast]
      summary: Projected cash balance combining manual entries and live pipeline
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200": { description: OK }
  /payments/{id}/reschedule:
    post:
      tags: [Payments]
      summary: Adjust a payment's due date (until it is posted); audit-logged
      description: Allowed while DRAFT, PENDING_APPROVAL or APPROVED. Moving an APPROVED payment to today or earlier posts it immediately.
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paymentDate]
              properties:
                paymentDate: { type: string, format: date }
                reason: { type: string }
      responses:
        "200": { description: OK }
        "409": { description: Payment already posted, cancelled or rejected }
...
  /forecasts/projection/detail:
    get:
      tags: [Forecast]
      summary: Day-by-day projection with per-bank/account breakdown
      description: >
        Simulated from today's balances. Per account per day: book, available (book - reserved - uncleared float)
        and liquidity (available + overdraft limit). Company totals are in the base currency unless currencyCode is given;
        currencies without an FX rate are listed in unconvertedCurrencies.
      parameters:
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
        - { in: query, name: currencyCode, schema: { type: string } }
        - { in: query, name: accountId, schema: { type: string } }
      responses:
        "200": { description: OK }
...
  /banker-acceptances:
    get:
      tags: [Banker Acceptances]
      summary: List banker acceptances (with tenor, discount and effective rate)
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [OUTSTANDING, SETTLED] } }
      responses:
        "200": { description: OK }
    post:
      tags: [Banker Acceptances]
      summary: Draw down a BA - credits the proceeds to the account (requires accounts.manage)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BankerAcceptanceInput" }
      responses:
        "201": { description: Created }
...
  /banker-acceptances/{id}/settle:
    post:
      tags: [Banker Acceptances]
      summary: Settle a BA at maturity - debits the settlement account (requires accounts.manage)
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                settledDate: { type: string, format: date }
                settledAmount: { type: number, description: "Defaults to the face amount" }
      responses:
        "200": { description: OK }
  /instrument-quotas:
    get:
      tags: [Quotas]
      summary: Cheque / bank draft daily quotas
      responses:
        "200": { description: OK }
    post:
      tags: [Quotas]
      summary: Create a quota (requires accounts.manage)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuotaInput" }
      responses:
        "201": { description: Created }
  /instrument-quotas/usage:
    get:
      tags: [Quotas]
      summary: Released / pending / remaining against each quota for a date
      parameters: [{ in: query, name: date, schema: { type: string, format: date } }]
      responses:
        "200": { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
    page:
      name: page
      in: query
      schema: { type: integer, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 20 }
    search:
      name: search
      in: query
      schema: { type: string }
  schemas:
...
    BankAccountInput:
      type: object
      required: [bankId, accountName, accountNumber, currencyCode]
      properties:
        bankId: { type: string }
        accountName: { type: string }
        accountNumber: { type: string }
        currencyCode: { type: string }
        accountType: { type: string, enum: [OPERATING, COLLECTION, DISBURSEMENT, RESERVE] }
        currentBalance: { type: number }
        minimumBalance: { type: number }
        targetBalance: { type: number }
        overdraftLimit: { type: number, description: "Approved overdraft; currentBalance may go negative down to -overdraftLimit" }
        siteName: { type: string, nullable: true, description: "Site/entity for the cash-reserve view, e.g. PJRM, Bukit Raja" }
    PaymentInput:
      type: object
      required: [beneficiaryName, amount, currencyCode, sourceAccountId, paymentDate]
      description: beneficiaryAccount and beneficiaryBank are required when paymentMethod is TRANSFER. currencyCode must equal the source account's currency.
      properties:
        beneficiaryName: { type: string }
        beneficiaryAccount: { type: string }
        beneficiaryBank: { type: string }
        amount: { type: number }
        currencyCode: { type: string }
        sourceAccountId: { type: string }
        paymentDate: { type: string, format: date, description: "Due date - the day the cash leaves the account" }
        paymentMethod: { type: string, enum: [TRANSFER, CHEQUE, BANK_DRAFT], default: TRANSFER }
        invoiceNumber: { type: string }
        description: { type: string }
    BankerAcceptanceInput:
      type: object
      required: [referenceNo, creditAccountId, faceAmount, proceedsAmount, drawdownDate, maturityDate]
      properties:
        referenceNo: { type: string }
        creditAccountId: { type: string }
        settlementAccountId: { type: string, description: "Defaults to the credited account; must be the same currency" }
        faceAmount: { type: number, description: "Payable at maturity" }
        proceedsAmount: { type: number, description: "Credited on drawdown; cannot exceed faceAmount" }
        drawdownDate: { type: string, format: date }
        maturityDate: { type: string, format: date }
        description: { type: string }
    QuotaInput:
      type: object
      required: [paymentMethod]
      description: At least one of dailyAmountLimit / dailyCountLimit is required.
      properties:
        paymentMethod: { type: string, enum: [CHEQUE, BANK_DRAFT] }
        bankId: { type: string, nullable: true, description: "null = all banks combined" }
        dailyAmountLimit: { type: number, nullable: true }
        dailyCountLimit: { type: integer, nullable: true }
        currencyCode: { type: string, default: MYR }
        isActive: { type: boolean, default: true }
...